resolve deps for go_repo to subrepo targets rather than its install (#108)
Previously we would add installs for go_repo. This means when we resolve deps we would add the install target as a dep when we could use the subrepo target directly. Using the subrepo target directly is more efficient since we need to build less and if we are using remote execution the data transfer is smaller.